+++ conan/tools/cmake/cmakedeps/__init__.py
@@ -1,0 +1,4 @@
+FIND_MODE_MODULE = "module"
+FIND_MODE_CONFIG = "config"
+FIND_MODE_NONE = "none"
+FIND_MODE_BOTH = "both"
