+++ conan/tools/cmake/cmakedeps/cmakedeps.py
@@ -1,6 +1,8 @@
 import os
 
 from conan.tools._check_build_profile import check_using_build_profile
+from conan.tools.cmake.cmakedeps import FIND_MODE_CONFIG, FIND_MODE_NONE, FIND_MODE_BOTH, \
+    FIND_MODE_MODULE
 from conan.tools.cmake.cmakedeps.templates.config import ConfigTemplate
 from conan.tools.cmake.cmakedeps.templates.config_version import ConfigVersionTemplate
 from conan.tools.cmake.cmakedeps.templates.macros import MacrosTemplate
@@ -10,11 +12,6 @@ from conan.tools.cmake.cmakedeps.templates.targets import TargetsTemplate
 from conans.errors import ConanException
 from conans.util.files import save
 
-
-FIND_MODE_MODULE = "module"
-FIND_MODE_CONFIG = "config"
-FIND_MODE_NONE = "none"
-FIND_MODE_BOTH = "both"
 
 
 class CMakeDeps(object):
